Add files via upload
This project is part of my personal healing journey. I started exploring Python Turtle graphics as a way to reconnect with my creativity and emotions after a difficult phase. I hope this inspires you to explore coding not just as a skill, but as a form of art and self-expression.